+++ resources/views/markdown/changelog.md
@@ -6,7 +6,8 @@ Travis Michael Cross <a href="mailto:travis@not.tv">travis@not.tv</a>
 
 ## v0.7.3
 * Fix Viewer Count not displaying correctly when watching a channel.
-* 
+* Fix Watch Now button on the Show page to play if any episode has a video or video_url.
+* Added a "The first episode is processing..." banner to the Show page if it has a video but the video is processing.
 
 ## v0.7.2
 * App loads bitchute and rumble videos from embed codes using ScraperAPI.